corrected the spawn height of ores
Just a small patch that should now spawn the ores at the correct heights. I noticed this when I saw lignite ore generate in a mountains biome at the height of 80 blocks, which was not what I had set in the default generation settings. So I found out that the maximum value when generating ores is also offsetted by the bottomOffset value. Then I made a secondary constructor for the OreGenerationSettings class that accepts IntRanges and automatically computes the offset of the maximum value. And this description got quite long...